Re-organize frontend components
