--- frontend/src/components/TableButton.module.css
@@ -1,3 +1,0 @@
-.items {
-  background-color: --color-tse-accent-blue-1;
-}
