--- frontend/src/components/TableButton.tsx
@@ -1,10 +1,0 @@
-import React from "react";
-import styles from "src/components/TableButton.module.css";
-
-export default function TableButton() {
-  return (
-    <div className={styles.items}>
-      <text>Button</text>
-    </div>
-  );
-}
