+++ frontend/src/components/shared/input/TextField/index.tsx
@@ -1,4 +1,4 @@
-import styles from "src/components/TextField.module.css";
+import styles from "@/components/shared/input/TextField/styles.module.css";
 import MUITextField, { TextFieldProps as MUITextFieldProps } from "@mui/material/TextField";
 import { ForwardedRef, forwardRef } from "react";
